+++ spec/wavefront-sdk/usage_spec.rb
@@ -2,7 +2,6 @@
 # frozen_string_literal: true
 
 require_relative '../spec_helper'
-require_relative '../test_mixins/general'
 
 # Unit tests for Usage class
 #