+++ HISTORY.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## 8.0.0 (2023-08-xx)
+## 8.0.0 (2023-09-05)
 * Drop support for Ruby 2.7 (Breaking change.)
 * Bump dependencies.
 * Add `Wavefront::AccessPolicy` class.
@@ -20,6 +20,7 @@
 * Add `Wavefront::ApiToken#describe_customer_token`.
 * Add `Wavefront::ApiToken#revoke_customer_token`, though it does nothing.
 * Add `Wavefront::MonitoredApplication` class.
+* Add `Wavefront::SpanSamplingPolicy` class.
 
 ## 7.0.0 (2023-01-19)
 * Drop support for Ruby 2.5. (Breaking change.)